## 1.1.0

### Release Highlights

- GraalVM Script engine support (https://www.graalvm.org/)
- support for case-sensitive database names and multi schema references (for supported databases)
- new MKDocs manual (https://www.benerator.de/ce/1.1.0/doc/)
- several new DemoCases to show how to use ... 
  - multi schema databases with references into foreign schemas 
  - case-sensitive database names like "cataloG"."Schema"."PersoN" 
  - GraalVM Script engine

### Important Notes

- Repository shift to github.com

### Breaking Changes

- removed Oracle Nashorn and replace it with GraalVM

+++ src/test/sql/create-tables.sql
@@ -24,10 +24,12 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-CREATE TABLE benerator.product (
-  id INTEGER UNSIGNED NOT NULL AUTO_INCREMENT,
-  name VARCHAR(45) NOT NULL,
-  category_id VARCHAR(5) NOT NULL,
-  price FLOAT(8,2) NOT NULL,
+CREATE TABLE benerator.product
+(
+    id          INTEGER UNSIGNED NOT NULL AUTO_INCREMENT,
+    name        VARCHAR(45) NOT NULL,
+    category_id VARCHAR(5)  NOT NULL,
+    price       FLOAT(8, 2
+) NOT NULL,
   PRIMARY KEY(`id`)
 );
